GCP Cloud Function1/requirements.txt
functions-framework==3.*
requests
pandas
google-cloud-bigquery==3.2.0
google-cloud-secret-manager==2.11.0
google-cloud-storage==2.14.0
Python-IO
db-dtypes
slack-sdk==3.21.3
tabulate==0.9.0
